LLM_service/requirements.txt
flask
openai==0.28
python-dotenv
llama-cpp-python
